--- src/Pipeline.hs
@@ -1,2 +1,0 @@
-module Pipeline where
-
